# DEV/DEV-03-openapi.yaml
openapi: 3.0.3
info:
  title: USTAPRO / SOZO Platform API
  version: 0.1.0
  description: |
    Контракт API платформы USTAPRO (Ташкент, ремонтно-сервисная платформа).
    Скелет для параллельной разработки фронтенда и бэкенда (монолит + PostgreSQL).

    Принципы:
    - Все изменения статуса заявки — только через POST /orders/{id}/transitions;
      недопустимый переход возвращает 409 с машиночитаемым кодом (PRD-05, раздел 1).
    - Суммы везде в тийинах (BIGINT, 1 сум = 100 тийин), поле `*_tiyin`.
    - Идемпотентность мобильных операций мастера — заголовок/поле `client_op_uuid`.
    - Аутентификация — JWT Bearer (access 15 мин / refresh 30 дней), вход по OTP-SMS.
    - Доменные события (order.*, billing.*, pricelist.release_activated) публикуются
      во внутреннюю шину; см. x-events в конце файла.
servers:
  - url: https://api.sozo.uz/v1
    description: Продакшн (домен sozo.uz — DEV-05 п.3; при смене бренда по итогам нейминга заменить единожды здесь)
  - url: https://staging-api.sozo.uz/v1
    description: Стейджинг
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      tags: [orders]
      operationId: createOrder
      summary: Создать заявку
      description: >
        B2C — от клиента, B2B — от сотрудника/руководителя точки.
        Дубль-детект: открытая заявка той же точки/адреса и категории за 24 ч
        возвращается предупреждением в ответе. Релиз прайса фиксируется по
        правилам ТЗ 3.7 (активный / релиз договора / релиз оценки дефекта).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заявка создана (статус new)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409':
          description: Долг блокирует создание / дубль требует подтверждения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [orders]
      operationId: listOrders
      summary: Список заявок в текущем контексте роли
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/OrderType'
        - name: location_id
          in: query
          schema: { type: string, format: uuid }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Страница заявок
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access-токен, полученный через /auth/verify (+ выбор контекста роли).
        В claims: user_id, role, organization_id?, location_id?, tenant_id.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 'Машиночитаемый код: домен/КОД (например order_state/INVALID_TRANSITION)'
          example: order_state/PHOTO_BEFORE_REQUIRED
        message:
          type: string
          description: Человекочитаемое сообщение на языке Accept-Language (рус/узб)
        details:
          type: object
          additionalProperties: true
          description: Контекст ошибки (текущий статус, допустимые действия и т.п.)
...
    OrderStatus:
      type: string
      description: Статусы заявки (ТЗ 4.1) + терминальные особые
      enum:
        - new
        - estimated
        - pending_approval
        - approved
        - assigned
        - master_departed
        - in_progress
        - addwork_approval
        - completed
        - verified
        - awaiting_payment
        - closed
        - rated
        - cancelled
        - dispute
      x-russian-label:
        new: 'Новая'
        estimated: 'Оценена'
        pending_approval: 'На утверждении'
        approved: 'Утверждена'
        assigned: 'Назначена'
        master_departed: 'Мастер выехал'
        in_progress: 'В работе'
        addwork_approval: 'Доп-согласование'
        completed: 'Выполнена'
        verified: 'Проверена'
        awaiting_payment: 'Ожидает оплаты'
        closed: 'Закрыта'
        rated: 'Оценена клиентом'
        cancelled: 'Отменена'
        dispute: 'Спор'
    OrderPause:
      type: string
      description: Паузы — суб-состояния «В работе» (не отдельные статусы)
      enum: [awaiting_materials, tech_break, blocked_third_party]
      x-russian-label:
        awaiting_materials: 'Ожидание материалов'
        tech_break: 'Технологический перерыв'
        blocked_third_party: 'Заблокировано третьей стороной'
    OrderType:
      type: string
      enum: [b2c, b2b, warranty, from_defect, inspection]
      x-russian-label:
        b2c: 'B2C'
        b2b: 'B2B'
        warranty: 'Гарантийная'
        from_defect: 'Из осмотра / дефекта'
        inspection: 'Осмотр плановый'
    TransitionAction:
      type: string
      description: >
        Действие статусной машины. Допустимость проверяется по графу типа
        заявки (PRD-05 разд. 1.4); недопустимое действие — 409.
      enum:
        - estimate
        - request_approval
        - approve
        - assign
        - depart
        - start
        - confirm_estimate
        - request_addwork
        - resolve_addwork
        - pause
        - resume
        - complete_stage
        - complete
        - verify
        - return_to_work
        - await_payment
        - close
        - rate
        - cancel
        - open_dispute
        - resolve_dispute
      x-russian-label:
        estimate: 'Оценить'
        request_approval: 'Отправить на утверждение'
        approve: 'Утвердить'
        assign: 'Назначить мастера'
        depart: 'Выехал'
        start: 'Начать работу (фото «до»)'
        confirm_estimate: 'Смета подтверждена клиентом'
        request_addwork: 'Запросить доп-работы'
        resolve_addwork: 'Решение по доп-смете'
        pause: 'Пауза'
        resume: 'Возобновить'
        complete_stage: 'Этап завершён'
        complete: 'Выполнена (фото «после»)'
        verify: 'Проверена диспетчером'
        return_to_work: 'Вернуть в работу'
        await_payment: 'В ожидание оплаты (санкция)'
        close: 'Закрыть (биллинг проведён)'
        rate: 'Оценить (клиент)'
        cancel: 'Отменить'
        open_dispute: 'Открыть спор'
        resolve_dispute: 'Разрешить спор'
...
    OrderCreate:
      type: object
      required: [type, category]
      properties:
        client_uuid:
          type: string
          format: uuid
          description: Идемпотентность создания (офлайн / ретраи)
        type:
          $ref: '#/components/schemas/OrderType'
        urgency:
          type: string
          enum: [normal, emergency]
          default: normal
        category: { type: string, example: 'santehnika' }
        location_id:
          type: string
          format: uuid
          description: B2B — точка; взаимоисключимо с address
        address:
          type: string
          description: B2C — адрес
        geo:
          type: object
          properties:
            lat: { type: number }
            lon: { type: number }
        description: { type: string }
        photo_ids:
          type: array
          maxItems: 5
          items: { type: string, format: uuid }
        desired_window:
          type: object
          properties:
            from: { type: string, format: date-time }
            to: { type: string, format: date-time }
        lines:
          type: array
          items:
            type: object
            properties:
              price_item_id: { type: string, format: uuid }
              qty: { type: number }
        defect_id:
          type: string
          format: uuid
          description: Для type=from_defect
        parent_order_id:
          type: string
          format: uuid
          description: Для type=warranty — исходная заявка
        promo_code: { type: string }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        number: { type: string, example: 'Z-2026-000123' }
        version: { type: integer, description: 'Для оптимистичной блокировки' }
        type:
          $ref: '#/components/schemas/OrderType'
        urgency:
          type: string
          enum: [normal, emergency]
        status:
          $ref: '#/components/schemas/OrderStatus'
        pause:
          allOf:
            - $ref: '#/components/schemas/OrderPause'
          nullable: true
        source:
          type: string
          enum: [app, phone, telegram]
        client_id: { type: string, format: uuid, nullable: true }
        location_id: { type: string, format: uuid, nullable: true }
        applicant_user_id:
          type: string
          format: uuid
          nullable: true
          description: Заявитель (программа баллов персонала точек)
        master_id: { type: string, format: uuid, nullable: true }
        price_list_release_id:
          type: string
          format: uuid
          description: Релиз прайса, зафиксированный заявкой (ТЗ 3.7)
        parent_order_id: { type: string, format: uuid, nullable: true }
        client_asset_id: { type: string, format: uuid, nullable: true }
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        quotes:
          type: array
          items:
            $ref: '#/components/schemas/Quote'
        assigned_window:
          type: object
          nullable: true
          properties:
            from: { type: string, format: date-time }
            to: { type: string, format: date-time }
        total_work_tiyin: { type: integer, format: int64 }
        total_material_tiyin: { type: integer, format: int64 }
        payment_channel:
          type: string
          enum: [subscription, invoice, online, cash]
          nullable: true
        warranty_days: { type: integer, example: 30 }
        allowed_actions:
          type: array
          description: Действия, допустимые из текущего статуса для текущей роли
          items:
            $ref: '#/components/schemas/TransitionAction'
        is_demo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    OrderLine:
      type: object
      description: Позиция заявки — копия цены на момент фиксации (иммутабельна)
      properties:
        id: { type: string, format: uuid }
        price_item_id: { type: string, format: uuid }
        name: { type: string, description: 'Копия названия позиции' }
        unit: { type: string, example: 'шт' }
        qty: { type: number }
        price_tiyin:
          type: integer
          format: int64
          description: Копия цены на момент заявки
        norm_hours: { type: number }
    Quote:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        kind:
          type: string
          enum: [initial, approved, additional, conservation]
          x-russian-label:
            initial: 'Первичная (вилка от–до)'
            approved: 'Утверждённая'
            additional: 'Доп-смета'
            conservation: 'Консервация'
        amount_tiyin: { type: integer, format: int64 }
        amount_to_tiyin:
          type: integer
          format: int64
          nullable: true
          description: Верх вилки для первичной B2C
        positions:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        photo_ids:
          type: array
          items: { type: string, format: uuid }
        approval_status:
          type: string
          enum: [pending, approved, declined, expired]
        approved_by: { type: string, format: uuid, nullable: true }
        approved_at: { type: string, format: date-time, nullable: true }
...